push onto github
this project is cloned from https://github.com/jianjianH/Order. since the 'storage' project from 《android studio开发实战》 is not stable. it always crashes when add new record. so, i find this project and test it well. so, store it onto my github for future reference. and will be modified to implement waybill functions.